Remove feature to look for duplicate cards
Remove feature to look for duplicate cards during card creation.
The feature was not working properly and is a bit redundant anyway.
From now on, the cards will be created in any case, which is not a problem, since the cards get tagged with the date of creation and are placed into a folder from which they can easily be removed if applicable.